+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # CHANGELOG
 ## [6.4.0-pre.8] - Unreleased
+### Added
+- Implemented octave shift annotation kinds "8va", "8vb" and "loco".
 
 ## [6.4.0-pre.7] - 2026-02-15
 ### Added